+++ util.h
@@ -9,12 +9,19 @@
 #ifdef _WIN32
 
 #include "windows.h"
+
 #else
 #include <cstdlib>
 #include <unistd.h>
 #endif
+
 void setWindowsConsoleToUTF8();
+
 void clearConsole();
-void setConsoleCursor(int i = 0, int j = 0);
-void calculateNewClosedCoord(int vectorSizeX, int vectorSizeY, int &x, int &y, int dirX, int dirY);
+
+void setConsoleCursor(short i = 0, short j = 0);
+
+void calculateNewClosedCoord(unsigned long vectorSizeX, unsigned long vectorSizeY, unsigned long &x, unsigned long &y,
+                             int dirX, int dirY);
+
 #endif //LIFE_UTIL_H
